# UserService/openapi/user.yaml
openapi: 3.0.3
info:
  title: User Service API
  version: 1.0.0
  description: Manages user registration, authentication, profile, MFA, and account recovery.

servers:
  - url: http://localhost:8101
    description: Local
  - url: https://user.foodly.example.com
    description: Production

tags:
  - name: Auth
  - name: Users

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserRegistrationRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '409': { description: Email already exists }

  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Auth tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthToken' }
        '401': { description: Invalid credentials }

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthToken' }
        '401': { description: Invalid refresh token }

  /users/me:
    get:
      tags: [Users]
      summary: Get current user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { description: Unauthorized }
    patch:
      tags: [Users]
      summary: Update current user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required: [id, email]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        phone: { type: string }
        role:
          type: string
          enum: [customer, admin, partner, courier]
        addresses:
          type: array
          items: { $ref: '#/components/schemas/Address' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    Address:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        country: { type: string }
        location:
          $ref: '#/components/schemas/GeoPoint'
      required: [line1, city, state, postalCode, country]

    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number, format: double }
        lng: { type: number, format: double }

    AuthToken:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer }

    UserRegistrationRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string }
        phone: { type: string }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }

    UserUpdateRequest:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        addresses:
          type: array
          items: { $ref: '#/components/schemas/Address' }

security:
  - bearerAuth: []